requirements.txt
pandas==2.2.3
boto3==1.35.79
duckdb==1.1.3
apache-airflow==2.10.3
matplotlib==3.8.0
seaborn==0.13.0
networkx==3.2.1
scikit-learn==1.3.1
